+++ kgx/transformers/logicterm_transformer.py
@@ -2,7 +2,7 @@ import logging
 import pandas as pd
 import networkx as nx
 from prologterms import Term, SExpressionRenderer, PrologRenderer
-from kgx.transformer import Transformer
+from kgx.transformers.transformer import Transformer
 
 from typing import Dict, List, Union
