+++ kgx/transformers/rdf_transformer.py
@@ -7,8 +7,8 @@ from collections import defaultdict
 from prefixcommons.curie_util import read_remote_jsonld_context
 
 from kgx.prefix_manager import PrefixManager
-from kgx.transformer import Transformer
-from kgx.rdf_graph_mixin import RdfGraphMixin
+from kgx.transformers.transformer import Transformer
+from kgx.transformers.rdf_graph_mixin import RdfGraphMixin
 from kgx.utils.rdf_utils import property_mapping, make_curie, infer_category
 from kgx.utils.kgx_utils import get_toolkit
